% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{lex}
\alias{lex}
\title{Break a string into labelled tokens based upon a set of patterns}
\usage{
lex(text, patterns, debug = FALSE)
}
\arguments{
\item{text}{a single character string}

\item{patterns}{a named vector of character strings.  Each string represents
a regex to match a token, and the name of the string is the
label for the token.  If the regex contains a captured group
it will be left as is, otherwise the whole regex will be
turned into a captured group. The patterns are used in order
such that an early match takes precedence over any later match.}

\item{debug}{print more debugging information about the matching. default: FALSE}
}
\value{
a named character vector with the names representing the token type
        and the contents representing the tokens
}
\description{
Break a string into labelled tokens based upon a set of patterns
}
\examples{
lex("hello there 123.45", patterns=c(number=pattern_number, word="\\\\w+", whitespace="\\\\s+"))

}
