% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{data}
\name{pattern_email}
\alias{pattern_email}
\title{Regex to match email addresses}
\format{An object of class \code{character} of length 1.}
\usage{
pattern_email
}
\description{
Regex to match email addresses
}
\keyword{datasets}
