% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{data}
\name{pattern_ipaddress}
\alias{pattern_ipaddress}
\title{Regex to match ip addresses}
\format{An object of class \code{character} of length 1.}
\usage{
pattern_ipaddress
}
\description{
Regex to match ip addresses
}
\keyword{datasets}
