% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\docType{data}
\name{pattern_number}
\alias{pattern_number}
\title{Regex to match numbers}
\format{An object of class \code{character} of length 1.}
\usage{
pattern_number
}
\description{
Regex to match numbers
}
\keyword{datasets}
