% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokenStream.R
\docType{data}
\name{TokenStream}
\alias{TokenStream}
\title{A class for manipulating/interrogating a stream of tokens}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TokenStream
}
\description{
A class for manipulating/interrogating a stream of tokens
}
\keyword{datasets}
